refactor(types): члены таблицы формы по её данным — в одном месте
`ТекущиеДанные` с `ДанныеСтроки` и тройка про идентификатор строки собирались
дважды одинаково: на типе вида данных и на типе конкретной таблицы. Вынесены в
`TableDataMembers` — оба вызова теперь читают один и тот же код.

Заодно закрыты замечания Sonar на этом PR: дублирующие зависимости ушли из
`FormItemTypesRegistrar` (S1200), в разборе реквизитов формы остался один выход
из цикла (S135), убран неиспользуемый импорт в тесте (S1128).
