+++ src/test/java/com/github/_1c_syntax/bsl/languageserver/types/registry/DynamicListTypesRegistrarTest.java
@@ -39,7 +39,6 @@ import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
-import static org.mockito.Mockito.when;
 
 /**
  * Разбор источника данных динамического списка
